Crate littlefs2_core

Core types for the littlefs2 crate.

See the documentation for littlefs2 for more information.

Macros

path

Creates a path from a string without a trailing null.

Structs

Ancestors

Iterator over the ancestors of a Path

Attribute

Custom user attribute that can be set on files and directories.

DirEntry

Error

The error type for filesystem operations.

FileOpenFlags

Definition of file open flags which can be mixed and matched as appropriate. These definitions are reminiscent of the ones defined by POSIX.

Iter

Iterator over the components of a Path

Metadata

File type (regular vs directory) and size of a file.

Path

A path

PathBuf

An owned, mutable path

Enums

FileType

Regular file vs directory

OpenSeekFrom

Enumeration of possible methods to seek within an file that was just opened Used in the read_chunk and write_chunk methods, Where SeekFrom::Current would not make sense.

PathError

Errors that arise from converting byte buffers into paths

SeekFrom

Enumeration of possible methods to seek within an I/O object.

Traits

DynFile

Object-safe trait for files.

DynFilesystem

Object-safe trait for filesystems.

Read

The Read trait allows for reading bytes from a file.

Seek

The Seek trait provides a cursor which can be moved within a file.

Vec

Write

The Write trait allows for writing bytes to a file.

Type Aliases

DirEntriesCallback

FileCallback

Predicate

Result