littlefs2_core
Macro path_buf

macro_rules! path_buf {
    ($path:literal) => { ... };
}

Creates an owned path from a string without a trailing null.

Panics and causes a compiler error if the string contains null bytes or non-ascii characters.

Examples

use littlefs2_core::{path_buf, PathBuf};

const HOME: PathBuf = path_buf!("/home");
let root = path_buf!("/");

Illegal values:

const WITH_NULL: PathBuf = path_buf!("/h\0me");  // does not compile

const WITH_UTF8: PathBuf = path_buf!("/höme");  // does not compile

The macro enforces const evaluation so that compilation fails for illegal values even if the macro is not used in a const context:

let path = path_buf!("te\0st");  // does not compile