littlefs2_core
Struct Error

pub struct Error { /* private fields */ }

The error type for filesystem operations.

Specific error codes are available as associated constants of this type.

assert_eq!(Error::IO.code(), -5);
assert_eq!(Error::new(-5), Some(Error::IO));

Implementations

impl Error

pub const IO: Self = _

Input / output error occurred.

pub const CORRUPTION: Self = _

File or filesystem was corrupt.

pub const NO_SUCH_ENTRY: Self = _

No entry found with that name.

pub const ENTRY_ALREADY_EXISTED: Self = _

File or directory already exists.

pub const PATH_NOT_DIR: Self = _

Path name is not a directory.

pub const PATH_IS_DIR: Self = _

Path specification is to a directory.

pub const DIR_NOT_EMPTY: Self = _

Directory was not empty.

pub const BAD_FILE_DESCRIPTOR: Self = _

Bad file descriptor.

pub const FILE_TOO_BIG: Self = _

File is too big.

pub const INVALID: Self = _

Incorrect value specified to function.

pub const NO_SPACE: Self = _

No space left available for operation.

pub const NO_MEMORY: Self = _

No memory available for completing request.

pub const NO_ATTRIBUTE: Self = _

No attribute or data available

pub const FILENAME_TOO_LONG: Self = _

Filename too long

pub const fn new(code: c_int) -> Option<Self>

Construct an Error from an error code.

Return values that are greater or equals to zero represent success. In this case, None is returned.

pub const fn code(&self) -> c_int

Return the error code of this error.